+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to semantic versioning.
 - Add `insertMany` function to the `Array` module.
 - Add `withIndex` function to the `Function` module.
 - Add `both` and `either` functions to the `Boolean` module.
+- Add `unsafeUnwrapLeft` function to the `Either` module.
 - Replace `exec` in the `String` module with `matchAll`.
 - Remove `contains` from the `Array` module as it's a duplicate of `elem` in fp-ts.
 - Rename `containsFlipped` in the `Array` module to `elemFlipped` for consistency with fp-ts.